requirements.txt
elevenlabs
requests
pyaudio
speechrecognition
vosk
pillow
vispy
numpy
torch
customtkinter
librosa
